@@ -1,9 +1,8 @@
 {
   "name": "baza",
-  "private": true,
   "version": "0.0.0",
+  "private": true,
   "type": "module",
-  "sourcemap": true,
   "scripts": {
     "dev": "cross-env NODE_ENV=development vite",
     "build-dev": "cross-env NODE_ENV=development vite build --mode development",
@@ -11,25 +10,20 @@
     "preview": "cross-env NODE_ENV=development vite preview"
   },
   "dependencies": {
-    "@vueuse/core": "^10.9.0",
-    "axios": "^1.6.7",
-    "floating-vue": "^5.2.2",
-    "jwt-decode": "^3.1.2",
-    "sass": "^1.70.0",
-    "vue": "^3.3.11",
-    "vue-jwt-decode": "^0.1.0",
-    "vue-router": "^4.2.5",
-    "vue-tippy": "^6.4.1",
+    "@vueuse/core": "^11.0.1",
+    "cross-env": "^7.0.3",
+    "env": "^0.0.2",
+    "vue": "^3.4.29",
+    "vue-inline-svg": "^3.1.3",
+    "vue-router": "^4.3.3",
     "vue-waypoint": "^4.3.0",
-    "vue3-dropper": "^0.1.6",
-    "vue3-lottie": "^3.2.0",
-    "vue3-recaptcha2": "^1.8.0",
-    "vuex": "^4.1.0",
-    "yandex-metrika-vue3": "^1.3.1"
+    "vue3-uuid": "^1.0.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^4.5.2",
-    "cross-env": "^7.0.3",
-    "vite": "^5.0.8"
+    "@vitejs/plugin-vue": "^5.0.5",
+    "axios": "^1.7.4",
+    "sass-embedded": "^1.77.8",
+    "uuid": "^10.0.0",
+    "vite": "^5.3.1"
   }
 }
